Replaced float with doubles for reduction

+++ Reduction/common.h
@@ -2,9 +2,9 @@
 #define _COMMON_H_
 
 //Can be changed to any associative and commutative function with an identity
-const float identity = 0.0f;
-__host__ __device__ float f(float a, float b);
+const double identity = 0.0;
+__host__ __device__ double f(double a, double b);
 
-float reduceGPU(float* input, unsigned int N, unsigned int type);
+double reduceGPU(double* input, unsigned int N, unsigned int type);
 
 #endif
