+++ Reduction/kernel.cu
@@ -12,16 +12,16 @@ void gpuAssert(cudaError_t code, const char *file, const int line) {
 	}
 }
 
-__host__ __device__ float f(float a, float b){
+__host__ __device__ double f(double a, double b){
 	return a+b;
 }
 
-__global__ void reduceKernel(float* input, float* partialSums, unsigned int N) {
+__global__ void reduceKernel(double* input, double* partialSums, unsigned int N) {
     unsigned int segment = 2 * blockDim.x * blockIdx.x;
     unsigned int i = segment + threadIdx.x;
 
     // Loading data to shared memory
-    __shared__ float input_s[BLOCK_DIM];
+    __shared__ double input_s[BLOCK_DIM];
     if (i + BLOCK_DIM < N) { input_s[threadIdx.x] = f(input[i], input[i + BLOCK_DIM]); }
     else if (i < N) { input_s[threadIdx.x] = input[i]; }
     else { input_s[threadIdx.x] = identity; }
@@ -41,13 +41,13 @@ __global__ void reduceKernel(float* input, float* partialSums, unsigned int N) {
 }
 
 // The higher the coarsening factor, the less we parallelize (which decreases the overhead for machines without enough resources)
-__global__ void reduceKernelWithThreadCoarsening(float* input, float* partialSums, unsigned int N) {
+__global__ void reduceKernelWithThreadCoarsening(double* input, double* partialSums, unsigned int N) {
     unsigned int segment = 2 * blockDim.x * blockIdx.x * COARSE_FACTOR;
     unsigned int i = segment + threadIdx.x;
 
     // Loading data to shared memory
-    __shared__ float input_s[BLOCK_DIM];
-	float sum = identity;
+    __shared__ double input_s[BLOCK_DIM];
+	double sum = identity;
 	for(unsigned int tile = 0; tile < 2 * COARSE_FACTOR && i + tile*BLOCK_DIM < N; tile++){
 		sum = f(sum, input[i+tile*BLOCK_DIM]);
 	}
@@ -67,14 +67,14 @@ __global__ void reduceKernelWithThreadCoarsening(float* input, float* partialSum
     }
 }
 
-float reduceGPU(float* input, unsigned int N, unsigned int type) {
+double reduceGPU(double* input, unsigned int N, unsigned int type) {
     Timer timer;
 
 	// Allocating GPU memory
     startTime(&timer);
 	
-    float *input_d;
-    cudaError_t errMallocA = cudaMalloc((void**) &input_d, N*sizeof(float)); cudaErrorCheck(errMallocA);
+    double *input_d;
+    cudaError_t errMallocA = cudaMalloc((void**) &input_d, N*sizeof(double)); cudaErrorCheck(errMallocA);
     
 	cudaDeviceSynchronize();
     stopTime(&timer);
@@ -83,7 +83,7 @@ float reduceGPU(float* input, unsigned int N, unsigned int type) {
     //Copying data to GPU from Host
     startTime(&timer);
     
-	cudaError_t errMemcpyA = cudaMemcpy(input_d, input, N*sizeof(float), cudaMemcpyHostToDevice); cudaErrorCheck(errMemcpyA);
+	cudaError_t errMemcpyA = cudaMemcpy(input_d, input, N*sizeof(double), cudaMemcpyHostToDevice); cudaErrorCheck(errMemcpyA);
 	
     cudaDeviceSynchronize();
     stopTime(&timer);
@@ -97,9 +97,9 @@ float reduceGPU(float* input, unsigned int N, unsigned int type) {
 	if (type != 1) { numElementsPerBlock *= COARSE_FACTOR; }
     const unsigned int numBlocks = (N + numElementsPerBlock - 1)/numElementsPerBlock;
 	
-    float* partialSums = (float*) malloc(numBlocks*sizeof(float));
-    float* partialSums_d;
-    cudaError_t errMallocB = cudaMalloc((void**) &partialSums_d, numBlocks*sizeof(float));  cudaErrorCheck(errMallocB);
+    double* partialSums = (double*) malloc(numBlocks*sizeof(double));
+    double* partialSums_d;
+    cudaError_t errMallocB = cudaMalloc((void**) &partialSums_d, numBlocks*sizeof(double));  cudaErrorCheck(errMallocB);
 	
     cudaDeviceSynchronize();
     stopTime(&timer);
@@ -116,7 +116,7 @@ float reduceGPU(float* input, unsigned int N, unsigned int type) {
 	//Copying data from GPU to Host
     startTime(&timer);
 
-    cudaError_t errMemcpyB = cudaMemcpy(partialSums, partialSums_d, numBlocks*sizeof(float), cudaMemcpyDeviceToHost); cudaErrorCheck(errMemcpyB);
+    cudaError_t errMemcpyB = cudaMemcpy(partialSums, partialSums_d, numBlocks*sizeof(double), cudaMemcpyDeviceToHost); cudaErrorCheck(errMemcpyB);
 	
     cudaDeviceSynchronize();
     stopTime(&timer);
@@ -125,7 +125,7 @@ float reduceGPU(float* input, unsigned int N, unsigned int type) {
     // Reducing partial sums on CPU
     startTime(&timer);
 	
-    float sum = identity;
+    double sum = identity;
     for(unsigned int i = 0; i < numBlocks; ++i) {
         sum = f(sum, partialSums[i]);
     }
